+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Added "Groups" page.
+- Added "Groups" page and group timelines.
+- Added ability to create posts in groups.
 - Added "Load conversation" item to post menu (admin only).
 
 ### Changed